Failed revert... made the app unbuildable.... Revert "Revert "First Stab at migrating many of the Double objects to Float's""
This reverts commit 31df9c86ed8b0466aa6c960accfff8856c4eacd6.

Conflicts:
	Modules/App/ColorGradients/ColorGradient.cs

+++ Common/Controls/ColorPicker/ColorPicker.cs
@@ -658,8 +658,8 @@ namespace Common.Controls.ColorManagement.ColorPicker
 				e.Handled = true;
 				return;
 			}
-			double value;
-			if (!double.TryParse(((TextBox) sender).Text,
+			float value;
+			if (!float.TryParse(((TextBox) sender).Text,
 			                     System.Globalization.NumberStyles.Integer,
 			                     null, out value)) return;
 
@@ -667,17 +667,17 @@ namespace Common.Controls.ColorManagement.ColorPicker
 
 			if (sender == tbHSV_H) {
 				HSV chsv = HSV.FromRGB(InternalColor.ToRGB());
-				chsv.H = value/360.0;
+				chsv.H = (float)value/360.0f;
 				InternalColor = XYZ.FromRGB(chsv.ToRGB());
 			}
 			else if (sender == tbHSV_S) {
 				HSV chsv = HSV.FromRGB(InternalColor.ToRGB());
-				chsv.S = value/100.0;
+				chsv.S = (float)value/100.0f;
 				InternalColor = XYZ.FromRGB(chsv.ToRGB());
 			}
 			else if (sender == tbHSV_V) {
 				HSV chsv = HSV.FromRGB(InternalColor.ToRGB());
-				chsv.V = value/100.0;
+				chsv.V = (float)value/100.0f;
 				InternalColor = XYZ.FromRGB(chsv.ToRGB());
 			}
 				#endregion
@@ -686,14 +686,14 @@ namespace Common.Controls.ColorManagement.ColorPicker
 			else if (_mode == Mode.HSV_RGB) {
 				RGB crgb = InternalColor.ToRGB();
 				if (sender == tbSecond_1) {
-					crgb.R = value/255.0;
+					crgb.R = value/255.0f;
 				}
 				else if (sender == tbSecond_2) {
-					crgb.G = value/255.0;
+					crgb.G = value/255.0f;
 				}
 				else //sender==tbSecond_3
 				{
-					crgb.B = value/255.0;
+					crgb.B = value/255.0f;
 				}
 				InternalColor = XYZ.FromRGB(crgb);
 			}
@@ -816,7 +816,7 @@ namespace Common.Controls.ColorManagement.ColorPicker
 			{
 				if (LockValue_V) {
 					HSV temp = HSV.FromRGB(value.ToRGB());
-					temp.V = 1.0;
+					temp.V = 1.0f;
 					value = XYZ.FromRGB(temp.ToRGB());
 				}
 				_color = value;
